Split site and api routes
